# perf(scanner)!: store and reuse `next_indent_length`