--- a/.editorconfig
+++ b/.editorconfig
@@ -1,0 +1,3 @@
+[scanner.c]
+indent_size = 4
+indent_style = space