@llllvvuu 2023-09-23
Required upstream change: https://github.com/tree-sitter/tree-sitter/pull/2648